+++ npc/cli/core.py
@@ -162,8 +162,10 @@ def _make_parser():
     parser_changeling.add_argument('name', help="Character name", metavar='name')
     parser_changeling.add_argument('seeming', help="The character's Seeming", metavar='seeming')
     parser_changeling.add_argument('kith', help="The character's Kith", metavar='kith')
-    parser_changeling.add_argument('-c', '--court', help="The character's Court", metavar='court')
-    parser_changeling.add_argument('-m', '--motley', help="The character's Motley", metavar='motley')
+    parser_changeling.add_argument('-c', '--court', help="Name of the character's Court", metavar='court')
+    parser_changeling.add_argument('-m', '--motley', help="Name of the character's Motley", metavar='motley')
+    parser_changeling.add_argument('-e', '--entitlement', help="Name of the character's entitlement", metavar='entitlement')
+    parser_changeling.add_argument('--freehold', help="Name of the character's freehold", metavar='freehold')
     parser_changeling.set_defaults(func=commands.create_changeling)
 
     # Subcommand for making changelings, with their unique options